pyproject.toml
[project]
name = "personalrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=0.7.700",
    "deepeval>=3.3.9",
    "llama-index-core~=0.12.41",
    "llama-index-embeddings-huggingface~=0.5.4",
    "llama-index-llms-openrouter~=0.3.2",
    "loguru>=0.7.3",
    "numpy<2",
    "python-dotenv>=1.1.1",
    "torch<=2.2.2",
    "torchaudio<=2.2.2",
    "torchvision<=0.17.2",
    "tqdm>=4.67.1",
    "websockets>=15.0.1",
]
